Extracts custom assertions to a separate file
